chainlit_app/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
supabase>=2.4.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0
reportlab>=4.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
